# docs/swagger.yaml
definitions:
...
  models.ApplicationCommand:
    properties:
      application_id:
        description: ID of the parent application
        type: string
      contexts:
        description: Interaction context(s) where the command can be used (global
          commands)
        items:
          type: string
        type: array
      default_member_permissions:
        description: Set of permissions represented as a bit set
        type: string
      default_permission:
        description: 'Deprecated: Indicates if the command is enabled by default when
          the app is added to a guild'
        type: boolean
      description:
        description: Description for CHAT_INPUT commands, 1-100 characters. Empty
          for USER and MESSAGE commands
        type: string
      description_localizations:
        additionalProperties:
          type: string
        description: Localization dictionary for description field
        type: object
      dm_permission:
        description: 'Deprecated: Indicates if the command is available in DMs for
          global commands'
        type: boolean
      guild_id:
        description: Guild ID of the command, if not global
        type: string
      handler:
        description: Determines whether the interaction is handled by the app's handler
          or by Discord
        type: string
      id:
        description: Unique ID of the command
        type: string
      integration_types:
        description: Installation contexts for the command (global commands)
        items:
          type: string
        type: array
      name:
        description: Name of the command, 1-32 characters
        type: string
      name_localizations:
        additionalProperties:
          type: string
        description: Localization dictionary for name field
        type: object
      nsfw:
        description: Indicates whether the command is age-restricted, defaults to
          false
        type: boolean
      options:
        description: Parameters for the command, max of 25 (CHAT_INPUT commands)
        items:
          $ref: '#/definitions/models.ApplicationCommandOption'
        type: array
      type:
        description: Type of command, defaults to 1
        type: integer
      version:
        description: Auto-incrementing version identifier updated during substantial
          record changes
        type: string
    type: object
  models.ApplicationCommandOption:
    properties:
      autocomplete:
        description: Whether autocomplete is enabled for this option
        type: boolean
      channel_types:
        description: Restrict shown channels to these types
        items:
          type: integer
        type: array
      choices:
        description: Optional array of choices for the user to pick from
        items:
          $ref: '#/definitions/models.ApplicationCommandOptionChoice'
        type: array
      description:
        description: 1-100 character description
        type: string
      description_localizations:
        additionalProperties:
          type: string
        description: Optional localization dictionary for the description field
        type: object
      max_length:
        description: Maximum allowed length (min 1, max 6000)
        type: integer
      max_value:
        description: Maximum value permitted (integer for INTEGER, double for NUMBER)
        type: number
      min_length:
        description: Minimum allowed length (min 0, max 6000)
        type: integer
      min_value:
        description: Minimum value permitted (integer for INTEGER, double for NUMBER)
        type: number
      name:
        description: 1-32 character name
        type: string
      name_localizations:
        additionalProperties:
          type: string
        description: Optional localization dictionary for the name field
        type: object
      options:
        description: Nested options if the option is a subcommand or subcommand group
        items:
          $ref: '#/definitions/models.ApplicationCommandOption'
        type: array
      required:
        description: Whether the option is required, default false
        type: boolean
      type:
        description: Type of the option
        type: integer
    type: object
  models.ApplicationCommandOptionChoice:
    properties:
      name:
        description: The name of the choice
        type: string
      value:
        description: The value of the choice (can be string, integer, or number)
    type: object
  models.AvatarDecorationData:
    properties:
      decoration:
        description: Example field for decoration
        type: string
    type: object
  models.Channel:
    properties:
      application_id:
        description: Optional application ID if bot-created group DM
        type: string
      applied_tags:
        description: Optional IDs of the tags applied to a thread in a forum/media
          channel
        items:
          type: string
        type: array
      available_tags:
        description: Optional set of tags available in a forum/media channel
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      bitrate:
        description: Optional bitrate (in bits) of the voice channel
        type: integer
      default_auto_archive_duration:
        description: Optional default duration (in minutes) for auto-archiving threads
        type: integer
      default_forum_layout:
        description: Optional default layout view for forum channels
        type: integer
      default_reaction_emoji:
        allOf:
        - $ref: '#/definitions/models.DefaultReaction'
        description: Optional default reaction emoji for threads
      default_sort_order:
        description: Optional default sort order for forum/media channels
        type: integer
      default_thread_rate_limit_per_user:
        description: Optional initial rate limit per user for newly created threads
        type: integer
      flags:
        description: Optional bitfield of channel flags
        type: integer
      guild_id:
        description: Optional Guild ID if the channel is part of a guild
        type: string
      icon:
        description: Optional icon hash for group DM
        type: string
      id:
        description: Snowflake ID of the channel
        type: string
      last_message_id:
        description: Optional ID of the last message sent in the channel
        type: string
      last_pin_timestamp:
        description: Optional timestamp of when the last pinned message was pinned
        type: string
      managed:
        description: Optional flag indicating if the group DM is managed by an application
        type: boolean
      member:
        allOf:
        - $ref: '#/definitions/models.ThreadMember'
        description: Optional thread member object for the current user
      member_count:
        description: Optional count of users in a thread
        type: integer
      message_count:
        description: Optional count of messages in a thread (excludes initial and
          deleted messages)
        type: integer
      name:
        description: Optional name of the channel (1-100 characters)
        type: string
      nsfw:
        description: Optional flag indicating if the channel is NSFW
        type: boolean
      owner_id:
        description: Optional owner ID for group DM or thread
        type: string
      parent_id:
        description: Optional ID of the parent category for guild channels
        type: string
      permission_overwrites:
        description: Optional explicit permission overwrites for members and roles
        items:
          $ref: '#/definitions/models.PermissionOverwrite'
        type: array
      permissions:
        description: Optional computed permissions for the invoking user in the channel
        type: string
      position:
        description: Optional sorting position of the channel
        type: integer
      rate_limit_per_user:
        description: Optional rate limit per user (in seconds)
        type: integer
      recipients:
        description: Optional list of recipients in a DM
        items:
          $ref: '#/definitions/models.User'
        type: array
      rtc_region:
        description: Optional voice region ID for the voice channel
        type: string
      thread_metadata:
        allOf:
        - $ref: '#/definitions/models.ThreadMetadata'
        description: Optional thread-specific fields
      topic:
        description: Optional topic of the channel (up to 4096 characters for forum/media
          channels, 1024 for others)
        type: string
      total_message_sent:
        description: Optional total number of messages ever sent in a thread
        type: integer
      type:
        description: Type of the channel
        type: integer
      user_limit:
        description: Optional user limit of the voice channel
        type: integer
      video_quality_mode:
        description: Optional video quality mode for the voice channel
        type: integer
    type: object
...
  models.DefaultReaction:
    properties:
      emoji_id:
        description: Snowflake ID of the emoji
        type: string
      emoji_name:
        description: Name of the emoji
        type: string
    type: object
...
  models.PermissionOverwrite:
    properties:
      allow:
        description: Allowed permissions bit set
        type: string
      deny:
        description: Denied permissions bit set
        type: string
      id:
        description: Snowflake ID of the overwrite (role or user)
        type: string
      type:
        description: Type of overwrite (0 = role, 1 = member)
        type: integer
    type: object
...
  models.Tag:
    properties:
      emoji:
        description: Optional emoji associated with the tag
        type: string
      id:
        description: Snowflake ID of the tag
        type: string
      name:
        description: Name of the tag
        type: string
    type: object
  models.ThreadMember:
    properties:
      flags:
        description: Thread member flags
        type: integer
      id:
        description: Snowflake ID of the thread member
        type: string
      join_timestamp:
        description: ISO8601 timestamp when the member joined
        type: string
      user_id:
        description: Snowflake ID of the user
        type: string
    type: object
  models.ThreadMetadata:
    properties:
      archive_timestamp:
        description: ISO8601 timestamp when the thread was archived
        type: string
      archived:
        description: Whether the thread is archived
        type: boolean
      auto_archive_duration:
        description: Duration in minutes to auto-archive
        type: integer
      locked:
        description: Whether the thread is locked
        type: boolean
    type: object
  models.User:
    properties:
      accent_color:
        description: Optional banner color as an integer
        type: integer
      avatar:
        description: Optional avatar hash of the user
        type: string
      avatar_decoration_data:
        allOf:
        - $ref: '#/definitions/models.AvatarDecorationData'
        description: Optional avatar decoration data
      banner:
        description: Optional banner hash of the user
        type: string
      bot:
        description: Optional flag indicating if the user is a bot
        type: boolean
      discriminator:
        description: User's Discord tag (four-digit identifier)
        type: string
      email:
        description: Optional user's email
        type: string
      flags:
        description: Optional flags on the user's account
        type: integer
      global_name:
        description: Optional display name (for bots, the application name)
        type: string
      id:
        description: Snowflake ID of the user
        type: string
      locale:
        description: Optional user's chosen language
        type: string
      mfa_enabled:
        description: Optional flag indicating if 2FA is enabled
        type: boolean
      premium_type:
        description: Optional Nitro subscription type
        type: integer
      public_flags:
        description: Optional public flags on the user's account
        type: integer
      system:
        description: Optional flag indicating if the user is a system user
        type: boolean
      username:
        description: Username of the user (not unique)
        type: string
      verified:
        description: Optional flag indicating if the email is verified
        type: boolean
    type: object
...
host: localhost:90
info:
  contact: {}
  description: API for managing Discord guilds using DiscordGo and Fiber.
  title: Discord Guild Management API
  version: "1.0"
paths:
...
  /api/guild/channels:
    get:
      description: Retrieve all channels from the guild.
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Channel'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Channels
      tags:
      - Channels
    post:
      description: Create a new channel in the guild.
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Channel'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create Guild Channel
      tags:
      - Channels
  /api/guild/channels/{channelid}:
    delete:
      description: Delete a specific channel in the guild.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Guild Channel
      tags:
      - Channels
    get:
      description: Retrieve a specific channel from the guild by ID.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Channel
      tags:
      - Channels
    patch:
      description: Update a specific channel in the guild.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update Guild Channel
      tags:
      - Channels
...
  /api/guild/channels/{channelid}/permissions/{overwriteid}:
    delete:
      description: Delete a specific permission overwrite for a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Overwrite ID
        in: path
        name: overwriteid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Channel Permissions
      tags:
      - Channels
    put:
      description: Edit permissions for a specific channel in the guild.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Overwrite ID
        in: path
        name: overwriteid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Edit Channel Permissions
      tags:
      - Channels
  /api/guild/commands:
    get:
      description: Retrieve all guild application commands.
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.ApplicationCommand'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Application Commands
      tags:
      - Commands
    post:
      description: Create a new guild application command.
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ApplicationCommand'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create Guild Application Command
      tags:
      - Commands
  /api/guild/commands/{cmdid}:
    delete:
      description: Delete a guild application command by ID.
      parameters:
      - description: Command ID
        in: path
        name: cmdid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Guild Application Command
      tags:
      - Commands
    get:
      description: Retrieve a specific guild application command by ID.
      parameters:
      - description: Command ID
        in: path
        name: cmdid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApplicationCommand'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Application Command
      tags:
      - Commands
...
swagger: "2.0"